iot_platform/requirements.txt
flask==3.0.3
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
paho-mqtt==2.1.0
requests==2.32.5
PyJWT==2.10.1

